cleanup generator-private methods
