+++ generators/app/generator.mjs
@@ -24,7 +24,7 @@ import BaseGenerator from '../base-application/index.mjs';
 import { checkNode } from './support/index.mjs';
 import gitOptions from '../git/options.mjs';
 import serverOptions from '../server/options.mjs';
-import { cleanupOldFiles, upgradeFiles } from '../cleanup.mjs';
+import { cleanupOldFiles } from '../cleanup.mjs';
 import prompts from './prompts.mjs';
 import { packageJson } from '../../lib/index.mjs';
 import statistics from '../statistics.cjs';
@@ -488,7 +488,6 @@ export default class JHipsterAppGenerator extends BaseGenerator {
     return this.asWritingTaskGroup({
       cleanup({ application }) {
         cleanupOldFiles(this, application);
-        upgradeFiles(this);
       },
     });
   }
